extract common methods
